기록 추가 구현 중

+++ app/src/main/java/com/example/buildupfrontend/retrofit/Request/ActivityRequest.kt
@@ -18,7 +18,7 @@ data class ActivityRequest(
 
 data class EditActivityRequest(
     val id: Long,
-    val categoryId: Long,
+    val categoryId: Int,
     val activityName: String,
     val hostName: String,
     val roleName: String,
